Refactored HTTP Requests, added current tracking

+++ PicoCode/WifiConnect.py
@@ -4,35 +4,53 @@ from time import sleep
 from picozero import pico_led
 import machine
 
-CONN_RESET_COUNT = 10
-TIMEOUT_COUNT = 4
+CONN_RESET_COUNT = 5
+TIMEOUT_COUNT = 2
 
 def led_blink():
+    '''Blinks the LED'''
     pico_led.on()
     sleep(0.1)
     pico_led.off()
     sleep(0.1)
 
-def SetupWifiInfo():
-    f = open('WifiInfo.txt', 'w')
-    # Used for testing, this should be taken from the app
-    print("Enter SSID and Password:")
-    ssid = input("SSID: ")
-    password = input("Password: ")
-    # Store the wifi login information locally to be used later
-    f.write(f"{ssid}, {password}")
-    f.close()
+def setupWifiInfo(useExisting):
+    ''' Determines whether wifi information is input or read from a file '''
+    if(useExisting):
+        f = open('WifiInfo.txt', 'r')
+        info = f.readline().split(',')
+        ssid = info[0]
+        password = info[1]
+        f.close()
+    else:
+        f = open('WifiInfo.txt', 'w')
+        # Used for testing, this should be taken from the app
+        print("Enter SSID and Password:")
+        ssid = input("SSID: ")
+        password = input("Password: ")
+        # Store the wifi login information locally to be used later
+        f.write(f"{ssid},{password}")
+        f.close()
+    print(f"Using SSID: {ssid} and Password: {password}")
     return ssid, password
 
 
-def APMode():
-    # Turns on AP mode in powersaving for the phone to access
+def APModeSetup(ip, ap):
+    '''Turns on AP mode in powersaving for the phone to access '''
     ap = network.WLAN(network.AP_IF)
     ap.config(ssid = "PlugPico", password = "Capstone", pm = network.WLAN.PM_POWERSAVE)
     ap.active(True)
+    ip = ap.ifconfig()[0]
+    return ip, ap
+
+def APModeDisconnect(ap):
+    ''' Turns off AP Mode '''
+    ap.active(False)
     return
 
-def connect(ssid, password):
+
+def connectWifi(ssid, password):
+    ''' Attempt to connect to the wifi network given the SSID and Password '''
     wlan = network.WLAN(network.STA_IF)
     timecount = 0
     # Retry connection until timeout reached
@@ -54,24 +72,7 @@ def connect(ssid, password):
     led_blink()
     led_blink()
     ip = wlan.ifconfig()[0]
+    
     print(f'Connected on {ip}')
     return ip
 
-def open_socket(ip):
-    # Open a socket
-    address = (ip, 80) # 80 <- Port Number
-    connection = socket.socket()
-    connection.bind(address)
-    connection.listen(1)
-    return connection
-
-def parseRequest(connection):
-    client = connection.accept()[0]
-    request = client.recv(1024)
-    request = str(request)
-    print(request.split()[1])
-    try:
-        request = request.split()[1] # Request has format http://{IP}/poweron?
-    except IndexError:
-        pass
-    return client, request
